Move internal/collector/processor/metrics to processor/
Also remove empty internal/collector/processor/ package.
